Move GunicornSafeAtoms to uvicorn.logging
